Move ItemTable types into single file
