+++ apps/web/components/ItemTable/columns.tsx
@@ -1,39 +1,10 @@
 import { Prisma } from '@gw2treasures/database';
-import { FunctionComponent, ReactNode } from 'react';
-import { db } from '@/lib/prisma';
+import { ReactNode } from 'react';
 import { EntityIcon } from '../Entity/EntityIcon';
 import { ItemLink } from '../Item/ItemLink';
 import { Rarity } from '../Item/Rarity';
 import { Coins } from '../Format/Coins';
-import { QueryModel } from './query';
-
-type Result<Select extends Prisma.ItemSelect> =
-  Awaited<ReturnType<typeof db.item.findFirstOrThrow<{ select: Select }>>>;
-
-export type OrderBy<T = Prisma.ItemOrderByWithRelationInput> = T | T[]
-
-export interface ItemTableColumn<Select extends Prisma.ItemSelect> {
-  id: GlobalColumnId,
-  order?: number,
-  select: Select,
-  align?: 'right',
-  orderBy?: [asc: OrderBy, desc: OrderBy]
-}
-
-export type ColumnModelTypes = {
-  'item': { select: Prisma.ItemSelect, orderBy: Prisma.ItemOrderByWithRelationInput },
-  'content': { select: Prisma.ContentSelect, orderBy: Prisma.ContentOrderByWithRelationInput },
-}
-
-export interface ExtraColumn<Id extends string, Model extends QueryModel, Select extends ColumnModelTypes[Model]['select']> {
-  id: Id,
-  select: Select,
-  title: string;
-  order?: number,
-  component: FunctionComponent<{ item: Result<Select & { id: true }> }>
-  align?: 'right',
-  orderBy?: [asc: OrderBy<ColumnModelTypes[Model]['orderBy']>, desc: OrderBy<ColumnModelTypes[Model]['orderBy']>]
-}
+import { ColumnModelTypes, ExtraColumn, GlobalColumnId, ItemTableColumn, QueryModel, Result } from './types';
 
 // typehelper
 function createColumn<Select extends Prisma.ItemSelect>(column: ItemTableColumn<Select>) {
@@ -42,8 +13,6 @@ function createColumn<Select extends Prisma.ItemSelect>(column: ItemTableColumn<
 export function extraColumn<Model extends QueryModel>(column: ExtraColumn<string, Model, ColumnModelTypes[Model]['select']>) {
   return column;
 }
-
-export type GlobalColumnId = 'id' | 'item' | 'icon' | 'name_de' | 'name_en' | 'name_es' | 'name_fr' | 'level' | 'rarity' | 'type' | 'vendorValue';
 
 export const globalColumnDefinitions = {
   id: createColumn({
